emis_backend/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.6.0
async-timeout==3.0.1
attrs==21.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.12.1
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-filebrowser-no-grappelli==4.0.1
django-filter==23.2
django-import-export==3.2.0
django-js-asset==2.0.0
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-uuslug==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.7
et-xmlfile==1.1.0
hiredis==1.1.0
hyperlink==21.0.0
idna==3.2
idna-ssl==1.1.0
incremental==17.5.0
isdayoff==1.0
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
msgpack==1.0.2
multidict==5.1.0
num2words==0.5.12
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.11
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.9.2
PyHamcrest==2.0.2
PyJWT==2.7.0
pyOpenSSL==20.0.1
python-docx==0.8.11
python-slugify==4.0.1
pytz==2021.1
PyYAML==6.0
requests==2.30.0
sentry-sdk==1.0.0
serpy==0.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
text-unidecode==1.3
Twisted==21.2.0
txaio==23.1.1
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==2.0.2
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zope.interface==6.0